third_party/blink/web_tests/external/wpt/css/css-values/hypot-pow-sqrt-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'pow(1,1)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'pow(1,1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(pow(1,1))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(pow(1,1))' should be valid in transform. got disallowed value ""
PASS 'pow(1,1)' as a computed value should serialize as '1'.
FAIL 'scale(pow(1,1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(pow(1,1))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'hypot(1)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'hypot(1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(hypot(1))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(hypot(1))' should be valid in transform. got disallowed value ""
PASS 'hypot(1)' as a computed value should serialize as '1'.
FAIL 'scale(hypot(1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(hypot(1))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'sqrt(1)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'sqrt(1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(sqrt(1))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(sqrt(1))' should be valid in transform. got disallowed value ""
PASS 'sqrt(1)' as a computed value should serialize as '1'.
FAIL 'scale(sqrt(1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(sqrt(1))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'calc(pow(1,1) - 0.5)' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(pow(1,1) - 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(pow(1,1) - 0.5))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(pow(1,1) - 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(pow(1,1) - 0.5)' as a computed value should serialize as '0.5'. assert_equals: 'calc(pow(1,1) - 0.5)' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(pow(1,1) - 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(pow(1,1) - 0.5))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(hypot(1) * 0.5)' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(hypot(1) * 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(hypot(1) * 0.5))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(hypot(1) * 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(hypot(1) * 0.5)' as a computed value should serialize as '0.5'. assert_equals: 'calc(hypot(1) * 0.5)' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(hypot(1) * 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(hypot(1) * 0.5))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(sqrt(1) - 1)' as a specified value should serialize as 'calc(0)'. assert_not_equals: 'calc(sqrt(1) - 1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sqrt(1) - 1))' as a specified value should serialize as 'scale(calc(0))'. assert_not_equals: 'scale(calc(sqrt(1) - 1))' should be valid in transform. got disallowed value ""
FAIL 'calc(sqrt(1) - 1)' as a computed value should serialize as '0'. assert_equals: 'calc(sqrt(1) - 1)' and '0' should serialize the same in computed values. expected "0" but got "1"
FAIL 'scale(calc(sqrt(1) - 1))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'. assert_equals: 'scale(calc(sqrt(1) - 1))' and 'matrix(0, 0, 0, 0, 0, 0)' should serialize the same in computed values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
Harness: the test ran to completion.
